requirements.txt
requests
python-csv
pyyaml
pymysql
cryptography